fix: darwin builds
